Register block render_callback for SSR in editor
Registers a block render_callback for this block. Because this
render_callback is only needed in the editor (the frontend markup is
handled using core's shortcode handler from the shortcode text stored in
the block content) this "callback" just checks for a sprecific attribute
which is only defined in the block's server-side registration.